+++ examples/V3_neuralfoil.jl
@@ -3,6 +3,7 @@ if Base.active_project() \!= joinpath(@__DIR__, "Project.toml")
     Pkg.activate(@__DIR__)
 end
 using GLMakie
+using MakieControlPlots
 using VortexStepMethod
 using LinearAlgebra