+++ examples/linearize_check.jl
@@ -4,6 +4,7 @@ if Base.active_project() != joinpath(@__DIR__, "Project.toml")
 end
 
 using GLMakie
+using MakieControlPlots
 using DifferentiationInterface
 using LinearAlgebra
 using VortexStepMethod